requirements.txt
numpy==1.26.4
pandas==2.2.2
scipy==1.12.0
pyyaml==6.0.1
pytest==7.4.4
